NetteExtension moved to Bridges
